Add json output to smartvio application
Adding json output to the SmartVIO application allows for easier
parsing in other tools. Note that in order to use the same json
library already in use by the dna-writer tool the smartvio
application has to be compiled as a C++ application, requiring
some modifications to how the SmartVIOConfig struct is initialized.
Much of the C code was kept consistent where possible.

There are also a few changes to clean things up and remove warnings.