Bugfix: Properly handle 16-bit current loads
The DNA generator previously clipped the top 8-bits of any current load
requiring the full two bytes. Since the data was lost when reading from
json rather than when writing to the DNA blob the CRC is calculated
from the incorrect data and will still pass when readin in by a SmartVIO
controller.